+++ stories/intro/Overview.stories.tsx
@@ -11,6 +11,7 @@ import itwinImage from '../../.storybook/public/itwin.svg';
 import itwinImageDark from '../../.storybook/public/itwin-dark.svg';
 import { Headline, Subheading, Title, Code } from '../../src/';
 import cx from 'classnames';
+import { CreeveyMeta } from 'creevey';
 
 const channel = addons.getChannel();
 
@@ -21,8 +22,12 @@ export default {
     options: { showPanel: false },
     controls: { disable: true },
     actions: { disable: true },
+    storysource: { disable: true },
+    creevey: {
+      skip: 'Overview',
+    },
   },
-} as Meta;
+} as Meta & CreeveyMeta;
 
 export const Overview: Story = () => {
   const [isDarkTheme, setIsDarkTheme] = React.useState(
@@ -30,18 +35,23 @@ export const Overview: Story = () => {
   );
 
   React.useEffect(() => {
-    const hideAddonsButton = parent.document.querySelector(
-      'button[title="Hide addons [A]"]',
+    const shortcutsButton = parent.document.querySelector(
+      'button[title="Shortcuts"]',
     ) as HTMLButtonElement;
+    shortcutsButton?.click();
 
-    const isPanelOpen = !!parent.document.querySelector(
-      '#storybook-preview-wrapper',
-    )?.parentElement?.parentElement?.style.height;
+    setTimeout(() => {
+      const addonsToggle = parent.document.querySelector('a#A') as HTMLElement;
+      const checkmark = addonsToggle?.firstElementChild?.firstElementChild;
 
-    if (isPanelOpen) {
-      hideAddonsButton.click();
-    }
-  });
+      // Panel is open if checkmark is an svg
+      if (checkmark?.tagName === 'svg') {
+        addonsToggle?.click();
+      } else {
+        shortcutsButton?.click();
+      }
+    });
+  }, []);
 
   React.useEffect(() => {
     const listener = (isDark: boolean) => setIsDarkTheme(isDark);
@@ -91,7 +101,7 @@ export const Overview: Story = () => {
           },
           img: {
             component: (args) =>
-              args.src.includes('iTwinUI_logo') ? (
+              args.src.includes('iTwinUI-logo') ? (
                 <img
                   src={isDarkTheme ? itwinImageDark : itwinImage}
                   width={167}